+++ PF-Rebirth/client/src/Components/Login/Login.css
@@ -62,30 +62,7 @@
   margin-top: 30px;
 }
 
-.btn-google {
-  width: 100%;
-  height: 50px;
-  padding: 12px 10px;
-  background-color: var(--grey) !important;
-  border: solid 1px var(--border);
-  color: var(--white);
-  border-radius: 6px;
-  text-align: center;
-  cursor: pointer;
-  margin: 40pt 0 40pt 0;
-  background: url("../../Assets/flat-color-icons_google.svg") no-repeat;
-  background-position: left 30% top 50%;
-  box-shadow: 0px 5px 5px 0px var(--grey2);
-}
 
-.btn-google:hover {
-  background-color: var(--greylight) !important;
-  border: solid 1px var(--border);
-  color: var(--white);
-  cursor: pointer;
-  background: url("../../Assets/flat-color-icons_google.svg") no-repeat;
-  background-position: left 30% top 50%;
-}
 
 .btn.btn-block {
   width: 100%;
