+++ PF-Rebirth/client/src/Tools/client_id.json
@@ -1,0 +1,1 @@
+{"web":{"client_id":"944240356620-nimmpjlb0om115mdrka0o1a1tsn8j6c9.apps.googleusercontent.com","project_id":"pruebarebirth","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"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","redirect_uris":["http://localhost:3000"],"javascript_origins":["http://localhost:3000"]}}
